requirements.txt
python-dotenv
pyarrow
pandas
scikit-learn 
pyodbc
fastparquet
sqlalchemy
psycopg2
pyyaml
openpyxl
matplotlib
seaborn
colormath
minisom